moved annotated model class
